+++ src/code42cli/cmds/auditlogs.py
@@ -4,14 +4,12 @@ from datetime import timezone
 
 import click
 
+import code42cli.options as opt
 from code42cli.click_ext.groups import OrderedGroup
 from code42cli.cmds.search.cursor_store import AuditLogCursorStore
-from code42cli.cmds.search.options import BeginOption
-from code42cli.date_helper import parse_max_timestamp
-from code42cli.date_helper import parse_min_timestamp
+from code42cli.date_helper import convert_datetime_to_timestamp
 from code42cli.logger import get_logger_for_server
-from code42cli.options import begin_option
-from code42cli.options import end_option
+from code42cli.options import checkpoint_option
 from code42cli.options import format_option
 from code42cli.options import sdk_options
 from code42cli.options import server_options
@@ -31,7 +29,14 @@ AUDIT_LOGS_DEFAULT_HEADER["actorIpAddress"] = "ActorIpAddress"
 AUDIT_LOGS_DEFAULT_HEADER["userName"] = "AffectedUser"
 AUDIT_LOGS_DEFAULT_HEADER["userId"] = "AffectedUserUID"
 
-
+begin_option = opt.begin_option(
+    AUDIT_LOGS_KEYWORD,
+    callback=lambda ctx, param, arg: convert_datetime_to_timestamp(arg),
+)
+end_option = opt.end_option(
+    AUDIT_LOGS_KEYWORD,
+    callback=lambda ctx, param, arg: convert_datetime_to_timestamp(arg),
+)
 filter_option_usernames = click.option(
     "--actor-username",
     required=False,
@@ -72,30 +77,15 @@ filter_option_event_types = click.option(
 
 
 def filter_options(f):
-    f = begin_option(
-        f,
-        AUDIT_LOGS_KEYWORD,
-        callback=lambda ctx, param, arg: parse_min_timestamp(arg),
-        cls=BeginOption,
-    )
-    f = end_option(
-        f, AUDIT_LOGS_KEYWORD, callback=lambda ctx, param, arg: parse_max_timestamp(arg)
-    )
     f = filter_option_event_types(f)
     f = filter_option_usernames(f)
     f = filter_option_user_ids(f)
     f = filter_option_user_ip_addresses(f)
     f = filter_option_affected_user_ids(f)
     f = filter_option_affected_usernames(f)
+    f = end_option(f)
+    f = begin_option(f)
     return f
-
-
-checkpoint_option = click.option(
-    "-c",
-    "--use-checkpoint",
-    metavar="checkpoint",
-    help="Only get audit-log events that were not previously retrieved.",
-)
 
 
 @click.group(cls=OrderedGroup)
@@ -117,7 +107,7 @@ def clear_checkpoint(state, checkpoint_name):
 @audit_logs.command()
 @filter_options
 @format_option
-@checkpoint_option
+@checkpoint_option(AUDIT_LOGS_KEYWORD)
 @sdk_options()
 def search(
     state,
@@ -170,7 +160,7 @@ def search(
 
 @audit_logs.command()
 @filter_options
-@checkpoint_option
+@checkpoint_option(AUDIT_LOGS_KEYWORD)
 @server_options
 @sdk_options()
 def send_to(