+++ CHANGELOG.md
@@ -27,6 +27,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - The error text when removing an employee from a detection list now references the employee
   by ID rather the username.
 
+- Improved help text for date option arguments.
+
 ## 1.1.0 - 2020-12-18
 
 ### Fixed